+++ modules/light-clients/09-localhost/codec.go
@@ -1,0 +1,16 @@
+package localhost
+
+import (
+	coreregistry "cosmossdk.io/core/registry"
+
+	"github.com/cosmos/ibc-go/v9/modules/core/exported"
+)
+
+// RegisterInterfaces registers the tendermint concrete client-related
+// implementations and interfaces.
+func RegisterInterfaces(registry coreregistry.InterfaceRegistrar) {
+	registry.RegisterImplementations(
+		(*exported.ClientState)(nil),
+		&ClientState{},
+	)
+}